Auth is now core feature of the app So moved it in Core/modules
