+++ src/app.module.ts
@@ -9,6 +9,7 @@ import configurations from './configs/configurations';
 import { DataSourcesOptions } from './configs/typeorm';
 import { ActivityLogsController } from './core/modules/activity-logs/activity-logs.controller';
 import { ActivityLogsModule } from './core/modules/activity-logs/activity-logs.module';
+import { AuthModule } from './core/modules/auth/auth.module';
 import { FilesModule } from './core/modules/files/files.module';
 import { MessengerModule } from './core/modules/messenger/messenger.module';
 import { WebhooksModule } from './core/modules/webhooks/webhooks.module';
@@ -20,7 +21,6 @@ import { AccounttypesController } from './modules/accounttypes/accounttypes.cont
 import { AccounttypesModule } from './modules/accounttypes/accounttypes.module';
 import { AttachmentsController } from './modules/attachments/attachments.controller';
 import { AttachmentsModule } from './modules/attachments/attachments.module';
-import { AuthModule } from './modules/auth/auth.module';
 import { ProvidersController } from './modules/providers/providers.controller';
 import { ProvidersModule } from './modules/providers/providers.module';
 import { RolesController } from './modules/roles/roles.controller';
