+++ src/core/modules/auth/auth.module.ts
@@ -1,9 +1,9 @@
 import { Logger, Module } from '@nestjs/common';
 import { JwtModule } from '@nestjs/jwt';
 import { PassportModule } from '@nestjs/passport';
-import { jwtModuleOptions } from '../../configs/jwt';
-import { ActivityLogsModule } from '../../core/modules/activity-logs/activity-logs.module';
-import { UsersModule } from '../users/users.module';
+import { jwtModuleOptions } from '../../../configs/jwt';
+import { UsersModule } from '../../../modules/users/users.module';
+import { ActivityLogsModule } from '../activity-logs/activity-logs.module';
 import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
 import { PassportJwtStrategy } from './strategies/jwt.strategy';
