+++ src/core/modules/auth/dtos/login.dto.ts
@@ -1,6 +1,6 @@
 import { ApiProperty, PickType } from '@nestjs/swagger';
 import { IsEmail, IsString, IsStrongPassword } from 'class-validator';
-import { User } from '../../users/entities/user.entity';
+import { User } from '../../../../modules/users/entities/user.entity';
 
 export class LoginRequestDto extends PickType(User, ['email', 'password']) {
   @ApiProperty({
